display, fetcher: Allow front history request to fail
This commit changes the fetcher such that if the request for a system's
front history fails, the result array is returned as-is (rather than
returning a fetcher fail), so that current fronters can be displayed.

This also changes the display controller so that the duration display
will correctly fall back to the current switch, rather than the awake
switch.